+++ CHANGES.md
@@ -251,6 +251,11 @@
 
     - Export `Minimize` type constructor.
 
+  * `XMonad.Actions.WorkspaceNames`
+
+    - Added `workspaceNamesListTransform` which makes workspace names visible
+      to external pagers.
+
   * Several `LayoutClass` instances now have an additional `Typeable`
     constraint which may break some advanced configs. The upside is that we
     can now add `Typeable` to `LayoutClass` in `XMonad.Core` and make it